tools/requirements.txt
python-telegram-bot>=20.0,<22.0
fastmcp>=0.1.0
flask>=3.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
trafilatura>=1.6.0
newspaper4k>=0.9.0
